Roller re-design
They will be trigger-collider circles that move upwards when detecting the ground.
There will be a limit to how much they can move upwards.
When moving upwards they will create a pressure value that will exert an upwards force on the body.
They will be affected by gravity as to fall downwards when there is no ground in the way.

The tank's forward movement could be constant or use an aggregate of the roller's pressure values to determine a traction value that creates forward movement.