Tank forward movement
I re-named the script so that we could add a pressure-based movement if we wanted to in the future.
The script just creates a force at the tank's forward vector; It will probably need dampening or limiting and an impulse force might be more suiting.

Also .normalized for gravity and this